feat((illustration)): created api route to get filesnames
