+++ src/config.ts
@@ -1,0 +1,3 @@
+const dev = process.env.NODE_ENV !== 'production'
+
+export const server = dev ? 'http://localhost:3000' : 'https://vamdigital.com'
